rag-system/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
openai>=1.0.0
pinecone-client
PyMuPDF
langchain
sqlalchemy
tenacity
psycopg2-binary
python-docx
scikit-learn
numpy
